Revert "Fixed a few warnings"
This reverts commit 521d7940a6b54d900e7be127d73a9a4caefbd040.

+++ libraries/StepperMotor/StepperMotor.cpp
@@ -26,12 +26,12 @@ void StepperMotor::oneStepCCW() {
 }
 
 void StepperMotor::stepCW(uint16_t microstepNb) {
-	for (uint16_t i=0; i<microstepNb; i++)
+	for (int i=0; i<microstepNb; i++)
 		oneStepCW();
 }
 
 void StepperMotor::stepCCW(uint16_t microstepNb) {
-	for (uint16_t i=0; i<microstepNb; i++)
+	for (int i=0; i<microstepNb; i++)
 		oneStepCCW();
 }
